refactor: rewrite Tevatron OmniEmbed as ST wrapper, rename file
Replace manual AbsEncoder with SentenceTransformerMultimodalEncoderWrapper.
Remove requires_package in favor of extra_requirements_groups.
Rename to tevatron_omni_embed_models.py.
